refactor(rpc): simplify error handling in gas estimators
