+++ src/rpc/estimation/utils.ts
@@ -1,8 +1,20 @@
-import { type Hex, decodeErrorResult, parseAbi, type StateOverride } from "viem"
-import { getAuthorizationStateOverrides } from "@alto/utils"
-import type { StateOverrides, UserOperationV06, UserOperationV07 } from "@alto/types"
+import {
+    type Hex,
+    decodeErrorResult,
+    parseAbi,
+    type StateOverride,
+    BaseError,
+    ContractFunctionRevertedError
+} from "viem"
+import { getAuthorizationStateOverrides, type Logger } from "@alto/utils"
+import type {
+    StateOverrides,
+    UserOperationV06,
+    UserOperationV07
+} from "@alto/types"
 import { toViemStateOverrides } from "../../utils/toViemStateOverrides"
 import type { AltoConfig } from "../../createConfig"
+import { ValidationErrors } from "@alto/types"
 
 export function parseFailedOpWithRevert(data: Hex) {
     try {
@@ -60,3 +72,83 @@ export function prepareStateOverride({
 
     return toViemStateOverrides(stateOverride)
 }
+
+export function decodeSimulateHandleOpError(
+    error: unknown,
+    logger: Logger
+): {
+    result: "failed"
+    data: string
+    code: number
+} {
+    // Check if it's a BaseError with ContractFunctionRevertedError
+    if (!(error instanceof BaseError)) {
+        return {
+            result: "failed",
+            data: "Unknown error, could not parse simulate validation result.",
+            code: ValidationErrors.SimulateValidation
+        }
+    }
+
+    const revertError = error.walk(
+        (e) => e instanceof ContractFunctionRevertedError
+    ) as ContractFunctionRevertedError
+
+    if (!revertError) {
+        return {
+            result: "failed",
+            data: "Unknown error, could not parse simulate validation result.",
+            code: ValidationErrors.SimulateValidation
+        }
+    }
+
+    if (!revertError.data?.args) {
+        logger.debug(
+            { err: error },
+            "ContractFunctionRevertedError has no args"
+        )
+        return {
+            result: "failed",
+            data: "Unknown error, could not parse simulate validation result.",
+            code: ValidationErrors.SimulateValidation
+        }
+    }
+
+    const errorName = revertError.data.errorName
+    const args = revertError.data.args
+
+    switch (errorName) {
+        case "FailedOp":
+            return {
+                result: "failed",
+                data: args[1] as string,
+                code: ValidationErrors.SimulateValidation
+            }
+
+        case "FailedOpWithRevert":
+            return {
+                result: "failed",
+                data: `${args[1]} ${parseFailedOpWithRevert(args[2] as Hex)}`,
+                code: ValidationErrors.SimulateValidation
+            }
+
+        case "CallPhaseReverted":
+            return {
+                result: "failed",
+                data: args[0] as Hex,
+                code: ValidationErrors.SimulateValidation
+            }
+
+        default:
+            logger.warn(
+                { errorName },
+                "Unknown ContractFunctionRevertedError name"
+            )
+            return {
+                result: "failed",
+                data: "Unknown error, could not parse simulate validation result.",
+                code: ValidationErrors.SimulateValidation
+            }
+    }
+}
+
